Extract apis constants (#6081)
* Extract Rest API common constants
